requirements.txt
asgiref==3.8.1
Django==5.0.6
django-tinymce==4.0.0
Paginator==0.5.1
pillow==10.3.0
six==1.16.0
sqlparse==0.5.0
tailwindcss==0.0.1
whitenoise==6.6.0
